moved Installation commands to their own namespace. Also cleaned InstallCommand
