+++ src/Commands/Installation/InstallationSettings.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace TCG\Voyager\Commands\Installation;
+
+class InstallationSettings
+{
+	public static routes() {
+		return "\n\nRoute::group(['prefix' => 'admin'], function () {\n    Voyager::routes();\n});\n";
+	}
+}
